requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
Django==4.2
django-widget-tweaks==1.4.12
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.174
llama-cpp-python==0.1.51
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openapi-schema-pydantic==1.2.4
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
pydantic==1.10.7
PyYAML==6.0
requests==2.30.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
tenacity==8.2.2
tomli==2.0.1
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==2.0.2
yarl==1.9.2
djangorestframework==3.14.0